test(react/table): update all table test cases to fit latest feature
